reflecting: cube: correction, perm4.tex
two orthagonal faces, as attached by a single edge cube,
permuted from one of the two opposing (orthagonal to the
rotating face) face, must be in the opposite position on
the second.
=========================================================
\e In summary, we have

$05^e=(F_2^e, D_2^e, B_4^e, U_2^e)^e$ and

$50_R^e=(R_e^1, R_e^4, R_e^3, R_e^2)$.

{\it (Just now we made a correction, here, $B_2^e\to B_4^e$; some
rationale, that may give a tell-tale sign of why this is
the way it must be.  $B$ and $F$ are opposing faces; We are
looking at a $R$ permuter.  The edge that is shared between
the front in position $F_2^e$ must be, once rotated by the
orthagonal face $R$, in the opposite position on the back
face $B_4^e$.)}
}% end of section 13
=========================================================
(looking back over past things, looking forward also)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>