orientation: topic-here
new topic: cube and cubelet and their orientation
Spin, we have called it in p/8/, the previous topic of
corners, edges, and centres.

We still approach the visual aspect, taking a step back
in order to progress with abstraction.

p/9, the view on this topic, perm4.tex still the track
to follow forward

(another view)

Signed-off-by: Gunter Liszewski (@S10) <gunter.liszewski@gmail.com>